Rename TP Python module supply_chain_processor
